+++ server/component/style/carousel/tpl_carousel.php
@@ -1,0 +1,7 @@
+<div id="<?php echo $this->id_prefix; ?>-carousel" class="carousel slide" data-ride="carousel">
+    <?php $this->output_indicator_wrapper(); ?>
+    <div class="carousel-inner">
+        <?php $this->output_carousel_items(); ?>
+    </div>
+    <?php $this->output_controls(); ?>
+</div>