+++ server/component/style/carousel/tpl_carousel_item.php
@@ -1,0 +1,3 @@
+<div class="carousel-item <?php echo $active; ?>">
+    <img class="d-block w-100" src="<?php echo $url; ?>" alt="<?php echo $alt; ?>">
+</div>