+++ server/component/style/carousel/tpl_control.php
@@ -1,0 +1,4 @@
+<a class="carousel-control-<?php echo $direction; ?>" href="#<?php echo $this->id_prefix; ?>-carousel" role="button" data-slide="<?php echo $direction; ?>">
+    <span class="fas fa-lg <?php echo $icon; ?>" aria-hidden="true"></span>
+    <span class="sr-only"><?php echo $direction; ?></span>
+</a>