+++ server/component/style/carousel/tpl_indicator.php
@@ -1,0 +1,1 @@
+<li data-target="#<?php echo $this->id_prefix; ?>-carousel" data-slide-to="<?php echo $i; ?>" class="<?php echo $active; ?>"></li>