+++ server/component/style/carousel/tpl_indicators.php
@@ -1,0 +1,3 @@
+<ol class="carousel-indicators">
+    <?php $this->output_indicators(); ?>
+</ol>